refactor: define Medusa URL only once
