+++ src/main.tsx
@@ -7,7 +7,8 @@ import "./assets/styles/global.css"
 import { AccountProvider } from "./context/account"
 import { CacheProvider } from "./context/cache"
 import { InterfaceProvider } from "./context/interface"
-import { medusaUrl, queryClient } from "./services/config"
+import { medusaUrl } from "./services/config"
+import queryClient from "./services/queryClient"
 import App from "./App"
 import { FeatureFlagProvider } from "./context/feature-flag"
 import { SteppedProvider } from "./components/molecules/modal/stepped-modal"
