+++ src/services/config.js
@@ -1,5 +1,3 @@
-import { QueryClient } from "react-query"
-
 let medusaUrl = "http://localhost:9000"
 
 // deprecated
@@ -12,14 +10,4 @@ if (process.env.GATSBY_MEDUSA_BACKEND_URL) {
   medusaUrl = process.env.GATSBY_MEDUSA_BACKEND_URL
 }
 
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      refetchOnWindowFocus: false,
-      staleTime: 90000,
-      retry: 1,
-    },
-  },
-})
-
-export { medusaUrl, queryClient }
+export { medusaUrl }
