+++ src/services/queryClient.js
@@ -1,0 +1,13 @@
+import { QueryClient } from "react-query"
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      refetchOnWindowFocus: false,
+      staleTime: 90000,
+      retry: 1,
+    },
+  },
+})
+
+export default queryClient
