+++ src/services/request.js
@@ -1,18 +1,7 @@
 import axios from "axios"
+import { medusaUrl } from "./config"
 
-let baseURL = "http://localhost:9000"
-
-// deprecated
-if (process.env.GATSBY_STORE_URL) {
-  baseURL = process.env.GATSBY_STORE_URL
-}
-
-// takes precedence over GATSBY_STORE_URL
-if (process.env.GATSBY_MEDUSA_BACKEND_URL) {
-  baseURL = process.env.GATSBY_MEDUSA_BACKEND_URL
-}
-
-const client = axios.create({ baseURL })
+const client = axios.create({ baseURL: medusaUrl })
 
 export default function medusaRequest(method, path = "", payload = {}) {
   const options = {
